move log_outputs to preprocess
This allows it to be used from both
preprocess and execute.

Fixes errors like

NameError: global name 'log_outputs' is not defined
